fixed getStatAlteration return type
Made the return type of getStatAlteration a float rather than a double

+++ Gremlin Gardens/Assets/Scripts/Food Testing/Food.cs
@@ -74,7 +74,7 @@ public class Food : Object
      * @param stat: the stat in question
      * @return: how much the given stat will change
      */
-    public double getStatAlteration(string stat)
+    public float getStatAlteration(string stat)
     {
         return alteredStats[stat];
     }
